refactor: split up payment form types and API types
